Adding debugging for nan

+++ magenta/models/cond_performancernn/models.py
@@ -136,7 +136,7 @@ class LSTMModel(BaseModel):
             if mode == 'train' or mode == 'eval':
               labels_flat = mg.common.flatten_maybe_padded_sequences(
                       labels, lengths)
-              labels_flat = tf.debugging.check_numerics(labels_flat, "labels_flat invalid")
+              labels_flat = tf.cast(tf.debugging.check_numerics(tf.cast(labels_flat, tf.float32), "labels_flat invalid"), tf.int64)
               logits_flat = tf.debugging.check_numerics(logits_flat, "logits_flat invalid")
               softmax_cross_entropy = tf.nn.sparse_softmax_cross_entropy_with_logits(
                     labels=labels_flat, logits=logits_flat)
